Add query attribute into @NamedQuery annotaion in Financial Class - AliiRad

+++ src/main/java/com/mftplus/storage/controller/api/UserApi.java
@@ -44,7 +44,7 @@ public class UserApi {
     @DELETE
     @Path("/delete/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response delete(@PathParam("id") int id){ // TODO remove int DataType and replace it with Long
+    public Response delete(@PathParam("id") Long id){ // TODO remove int DataType and replace it with Long
         try {
             return Response.ok().entity(userService.remove(id)).build();
         } catch (Exception e) {
@@ -71,7 +71,7 @@ public class UserApi {
     @GET()
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/findById/{id}")
-    public Response findById(@PathParam("id") int id) { // TODO remove int DataType and replace it with Long
+    public Response findById(@PathParam("id") Long id) { // TODO remove int DataType and replace it with Long
         System.out.println("FindById Method in UserApi");
         System.out.println("Id : "+id);
         try {
